Fix Header component and client's history/invoice pages
All pages with header component had the same extra style attached as a temporary fix. Fixed header component so that the style was no longer necessary to fix it.

Also fixed the style of the client's history/invoice pages to the current app style, as it was revealed that it hadn't been adjusted yet.